+++ src/main/java/com/example/housemoney2/controller/PaymentsController.java
@@ -1,0 +1,5 @@
+package com.example.housemoney2.controller;
+
+public class PaymentsController {
+
+}
