+++ src/main/java/com/example/housemoney2/controller/PostsController.java
@@ -1,0 +1,5 @@
+package com.example.housemoney2.controller;
+
+public class PostsController {
+
+}
