+++ src/main/java/com/example/housemoney2/entity/User.java
@@ -1,0 +1,5 @@
+package com.example.housemoney2.entity;
+
+public class User {
+
+}
