+++ src/main/java/com/example/housemoney2/form/CreateUserForm.java
@@ -1,0 +1,5 @@
+package com.example.housemoney2.form;
+
+public class CreateUserForm {
+
+}
