+++ src/main/java/com/example/housemoney2/mapper/UserMapper.java
@@ -1,0 +1,5 @@
+package com.example.housemoney2.mapper;
+
+public interface UserMapper {
+
+}
