+++ src/main/java/com/example/housemoney2/service/CreateUserService.java
@@ -1,0 +1,5 @@
+package com.example.housemoney2.service;
+
+public class CreateUserService {
+
+}
