+++ src/main/java/com/example/housemoney2/service/LoginUserService.java
@@ -1,0 +1,5 @@
+package com.example.housemoney2.service;
+
+public class LoginUserService {
+
+}
