--- src/main/java/controller/PaymentsController.java
@@ -1,5 +1,0 @@
-package controller;
-
-public class PaymentsController {
-
-}
