--- src/main/java/controller/PostsController.java
@@ -1,5 +1,0 @@
-package controller;
-
-public class PostsController {
-
-}
