--- src/main/java/form/CreateUserForm.java
@@ -1,5 +1,0 @@
-package form;
-
-public class CreateUserForm {
-
-}
