--- src/main/java/form/LoginUserForm.java
@@ -1,5 +1,0 @@
-package form;
-
-public class LoginUserForm {
-
-}
