--- src/main/java/mapper/UserMapper.java
@@ -1,5 +1,0 @@
-package mapper;
-
-public interface UserMapper {
-
-}
