--- src/main/java/service/CreateUserService.java
@@ -1,5 +1,0 @@
-package service;
-
-public class CreateUserService {
-
-}
