--- src/main/java/service/LoginUserService.java
@@ -1,5 +1,0 @@
-package service;
-
-public class LoginUserService {
-
-}
